+++ entrypoint.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+# Exit on error
+set -e
+
+# Run the command
+exec npm start