+++ src/ui/ChangeLog.md
@@ -1,8 +1,11 @@
-## V1.0.0 (Mon July 1 2024)
+## V1.0.0 (Mon July 16 2024)
 
 ✨ New Feature
 
 - Add Button Story
+- Add Title Story
+- Add Input Story
+- Add Textarea Story
 
 ---